# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "INFRACTION-COMMUNE-BACKEND API",
        "description": "API for managing infractions, communes, decisions, and users in the commune system",
        "contact": {
            "name": "API Support",
            "email": "support@infraction-commune.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/decision": {
            "get": {
                "tags": [
                    "Decisions"
                ],
                "summary": "Get all decisions",
                "description": "Retrieve a list of all decisions in the system",
                "operationId": "getDecisions",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2023-12-15"
                                            },
                                            "decisionprise": {
                                                "type": "string",
                                                "example": "Fine of 150 DH"
                                            },
                                            "infraction_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Decisions"
                ],
                "summary": "Create a new decision",
                "description": "Create a new decision for an infraction",
                "operationId": "createDecision",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "decisionprise",
                                    "infraction_id"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-15"
                                    },
                                    "decisionprise": {
                                        "type": "string",
                                        "maxLength": 200,
                                        "minLength": 5,
                                        "example": "Fine of 150 DH"
                                    },
                                    "infraction_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Decision created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Decision created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-12-15"
                                                },
                                                "decisionprise": {
                                                    "type": "string",
                                                    "example": "Fine of 150 DH"
                                                },
                                                "infraction_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
...
    },
...
}